backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
yfinance==0.2.50
pandas==2.2.3
numpy==2.1.3
anthropic==0.40.0
python-dotenv==1.0.1
httpx==0.27.2
